requirements.txt
streamlit
openai
python-dotenv
gtts